# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AIStatus {
  available: boolean;
}

export type AIChatMessageRole =
  (typeof AIChatMessageRole)[keyof typeof AIChatMessageRole];

export const AIChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface AIChatMessage {
  role: AIChatMessageRole;
  /** @maxLength 4000 */
  content: string;
}

export interface AIChatInput {
  /** @maxItems 10 */
  messages: AIChatMessage[];
  /**
   * @minLength 1
   * @maxLength 4000
   */
  userMessage: string;
}

export interface AIChatOutput {
  response: string;
  /** Whether the latest user message expresses sadness or emotional pain. */
  sadnessDetected: boolean;
  /** Whether the latest user message indicates self-harm or immediate safety risk. */
  crisisDetected: boolean;
}

export interface AnalyzeEmotionInput {
  /**
   * @minLength 1
   * @maxLength 4000
   */
  text: string;
  /** @maxItems 12 */
  emojis: string[];
}

export type EmotionAnalysisSentiment =
  (typeof EmotionAnalysisSentiment)[keyof typeof EmotionAnalysisSentiment];

export const EmotionAnalysisSentiment = {
  positive: "positive",
  negative: "negative",
  neutral: "neutral",
} as const;

export type EmotionAnalysisCrisisRisk =
  (typeof EmotionAnalysisCrisisRisk)[keyof typeof EmotionAnalysisCrisisRisk];

export const EmotionAnalysisCrisisRisk = {
  none: "none",
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface EmotionAnalysis {
  emotion: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  intensity: number;
  sentiment: EmotionAnalysisSentiment;
  crisisRisk: EmotionAnalysisCrisisRisk;
  suggestions: string[];
  summary: string;
}

export interface AICheerOutput {
  message: string;
}