# lib/api-zod/src/generated/types/aIChatInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AIChatMessage } from "./aIChatMessage";

export interface AIChatInput {
  /** @maxItems 10 */
  messages: AIChatMessage[];
  /**
   * @minLength 1
   * @maxLength 4000
   */
  userMessage: string;
}