# lib/api-zod/src/generated/types/aIChatMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AIChatMessageRole } from "./aIChatMessageRole";

export interface AIChatMessage {
  role: AIChatMessageRole;
  /** @maxLength 4000 */
  content: string;
}