# lib/api-zod/src/generated/types/aIChatOutput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AIChatOutput {
  response: string;
  /** Whether the latest user message expresses sadness or emotional pain. */
  sadnessDetected: boolean;
  /** Whether the latest user message indicates self-harm or immediate safety risk. */
  crisisDetected: boolean;
}