# lib/api-zod/src/generated/types/analyzeEmotionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeEmotionInput {
  /**
   * @minLength 1
   * @maxLength 4000
   */
  text: string;
  /** @maxItems 12 */
  emojis: string[];
}