# lib/api-zod/src/generated/types/emotionAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmotionAnalysisCrisisRisk } from "./emotionAnalysisCrisisRisk";
import type { EmotionAnalysisSentiment } from "./emotionAnalysisSentiment";

export interface EmotionAnalysis {
  emotion: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  intensity: number;
  sentiment: EmotionAnalysisSentiment;
  crisisRisk: EmotionAnalysisCrisisRisk;
  suggestions: string[];
  summary: string;
}